pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "filekor"
version = "0.1.1"
description = "PDF metadata extraction CLI using PyExifTool"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "filekor maintainer"}
]
dependencies = [
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pyexiftool>=0.5.0",
    "pyyaml>=6.0.0",
    "pypdf>=6.9.1",
    "google-genai>=0.3.0",
    "rich>=13.0.0",
    "openai>=1.0.0",
]

[project.scripts]
filekor = "filekor.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
include = ["filekor*"]

[dependency-groups]
dev = [
    "pdbpp>=0.12.1",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
]
